make all packets inherit from BasePacket
