+++ try1/Core/Networking/Packets/BasePacket.cs
@@ -14,36 +14,7 @@ namespace ChatChimpClient.Core.Networking.Packets
         {
             MemoryStream ms = new MemoryStream(data);
             reader = new BinaryReader(ms);
-            writer = new BinaryWriter(ms);
         }
-        
-        // ---------------- WRITER FUNCTIONS ----------------
-
-        public void writeByte(byte b)
-            => writer.Write(b);
-
-        public void writeUshort(ushort number)
-        {
-            byte[] toWrite = BitConverter.GetBytes(number);
-            for ( int x = 0; x < 2; x++ )
-            {
-                writer.Write(toWrite[x]);
-            }
-        }
-
-        public void writeInt(int number)
-        {
-            writer.Write7BitEncodedInt(number);
-        }
-
-        public void writeString(string letters)
-        {
-            int msgLen = Encoding.UTF8.GetBytes(letters).Length;
-            writeInt(msgLen);
-            writer.Write(letters);
-        }
-
-        // ------------ READING FUNCTIONS --------------
 
         public int readIntBytes()
         {
