+++ try1/Core/Networking/Packets/InitPacket.cs
@@ -7,10 +7,10 @@ using ChatChimpClient.Core.PacketHandlers;
 
 namespace ChatChimpClient.Core.Networking.Packets
 {
-    public class initPacket
+    public class initPacket : BasePacket
     {
         public int response { get; set; }
-        public initPacket(PackageReader reader)
+        public initPacket(byte[] data) : base(data)
         {
             response = reader.readInt();
         }
